you save value

+++ app/helpers/dashboards_helper.rb
@@ -37,7 +37,7 @@ module DashboardsHelper
 					allowed_best_deal=category_best_deal(deal_type,sp,zip_code,1,false)
 					if allowed_best_deal.present?
 						if allowed_best_deal.effective_price.to_f>0
-							you_save = 12*(app_user_current_plan - allowed_best_deal.effective_price.to_f)
+							you_save = 12*(app_user_current_plan - allowed_best_deal.effective_price.to_i)
 						else
 							you_save = 12*(app_user_current_plan - allowed_best_deal.price)
 						end
@@ -83,7 +83,7 @@ module DashboardsHelper
 
 					if allowed_best_deal.present?
 						if allowed_best_deal.effective_price.to_f>0
-							you_save =  12*(app_user_current_plan - allowed_best_deal.effective_price.to_f)
+							you_save =  12*(app_user_current_plan - allowed_best_deal.effective_price.to_i)
 						else
 							you_save =  12*(app_user_current_plan - allowed_best_deal.price)
 						end
